// src-backup/types/user.ts
/*
|--------------------------------------------------------------------------
| User
|--------------------------------------------------------------------------
*/

export interface User {

    /*
    |--------------------------------------------------------------------------
    | Identity
    |--------------------------------------------------------------------------
    */

    id?: string;

    client_id: string;

    /*
    |--------------------------------------------------------------------------
    | User Information
    |--------------------------------------------------------------------------
    */

    name: string;

    email: string;

    /*
    |--------------------------------------------------------------------------
    | Authentication
    |--------------------------------------------------------------------------
    */

    password_hash: string;

    /*
    |--------------------------------------------------------------------------
    | Authorization
    |--------------------------------------------------------------------------
    */

    role: string;

    /*
    |--------------------------------------------------------------------------
    | Timestamps
    |--------------------------------------------------------------------------
    */

    created_at?: string;

    updated_at?: string;

}